Added more functionality
Added the keypress event so the user can press enter to search and to submit items to the store. I made it so the search function displays only the items search. Made the categories section on the side so the user can click on them and search for specific item types. I also made the categories static instead of letting them update based on what was being added.